Split admin notices into a general Notices and a removable Premium component
